requirements.txt
# Backend dependencies
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
matplotlib==3.7.2
opencv-python==4.8.0.76
pymongo==4.5.0
pyjwt==2.8.0
cryptography==41.0.3
werkzeug==2.3.7

# Data processing
networkx==3.1
nltk==3.8.1
spacy==3.6.1

# Testing
pytest==7.4.0
